Extract persistence test utility methods to the separate object
